Move room data collection into its own class
